tests: Add debug console tests
Add initial tests for the debug console feature. This required changing the
lazy static vector of shells to use an arc mutex (so it's thread-safe and
mutable). The vector is also initialised to an empty list for the tests, but
set to the standard set of shells for release builds.

Signed-off-by: James O. D. Hunt <james.o.hunt@intel.com>